refactoring path aliases pt.2
