+++ src/pages/user/CompleteProfile.page.tsx
@@ -1,11 +1,12 @@
 import { FormEventHandler, useState } from "react";
 import { Navigate } from "react-router-dom";
 import { ErrorAlert } from "@components";
-import { type UserProfile, createUserProfile } from "@services/utils";
+import { createUserProfile } from "@/services/utils";
 import { useAuth, useNotification } from "@providers";
 import { routes } from "@/navigation/constants";
-import { Profile, formValidationSchema } from "../../components/forms/Profile";
-import { defaultProfile } from "../../components/forms/defaults";
+import { Profile, formValidationSchema } from "@/components/forms/Profile";
+import { defaultProfile } from "@/components/forms/defaults";
+import type { UserProfile } from "@/services/utils/types";
 
 // TODO: refactor imports
