+++ src/providers/types.ts
@@ -1,1 +1,2 @@
 export type { ProviderName } from "./auth.provider";
+export type { NotificationOptions } from "./notification.provider";
