--- src/services/index.ts
@@ -1,6 +1,0 @@
-/*
- *
- * @service Firebase 
- * 
- */
-export { app, analytics, firestore, auth } from "./firebase";
