+++ src/services/utils/types.ts
@@ -1,0 +1,19 @@
+export interface UserTicketData {
+    userId: string;
+    firstName: string;
+    lastName: string;
+}
+
+export interface UserProfile {
+    id: string;
+    firstName: string;
+    lastName: string;
+    email: string;
+    emailVerified: boolean;
+    phone: string;
+    school: string;
+    levelOfStudy: string;
+    countryOfResidence: string;
+    age: string;
+    discord: string;
+}
